+++ modules/spring-web/src/main/kotlin/com/explyt/spring/web/builder/openapi/yaml/OpenApiYamlPathBuilder.kt
@@ -18,21 +18,28 @@
 package com.explyt.spring.web.builder.openapi.yaml
 
 import com.explyt.spring.web.builder.openapi.OpenApiPathBuilder
-import com.explyt.spring.web.inspections.quickfix.AddEndpointToOpenApiIntention
+import com.explyt.spring.web.inspections.quickfix.AddEndpointToOpenApiIntention.EndpointInfo
 
 class OpenApiYamlPathBuilder(
-    private val endpoint: AddEndpointToOpenApiIntention.EndpointInfo,
+    private val path: String,
     indent: String = "",
     builder: StringBuilder = StringBuilder()
 ) : OpenApiPathBuilder(indent, builder), YamlKeyValueGenerator {
 
     override fun build() {
         builder.appendLine()
-        builder.append("$indent${endpoint.path}:")
-        for (httpType in endpoint.requestMethods) {
-            OpenApiYamlPathHttpTypeBuilder(endpoint, httpType, "$indent  ", builder)
-                .build()
+        builder.append("$indent${path}:")
+        for (endpoint in endpointInfos) {
+            for (httpType in endpoint.requestMethods) {
+                OpenApiYamlPathHttpTypeBuilder(endpoint, httpType, "$indent  ", builder)
+                    .build()
+            }
         }
+    }
+
+    override fun addEndpoint(endpoint: EndpointInfo): OpenApiYamlPathBuilder {
+        super.addEndpoint(endpoint)
+        return this
     }
 
 }